+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All Notable changes to `League\Uri\Components` will be documented in this file
 
+## 1.1.1 - 2017-11-03
+
+### Added
+
+- `League\Uri\parse_query` as an alias of `Query::parse`
+- `League\Uri\extract_query` as an alias of `Query::extract`
+
+### Fixed
+
+- `League\Uri\parse_query` returned value was the wrong one
+
+### Deprecated
+
+- None
+
+### Remove
+
+- None
+
 ## 1.1.0 - 2017-10-24
 
 ### Added